// web/src/api/generated/models/SessionView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AccountView } from './AccountView';
import {
    AccountViewFromJSON,
    AccountViewFromJSONTyped,
    AccountViewToJSON,
    AccountViewToJSONTyped,
} from './AccountView';
import type { TokenView } from './TokenView';
import {
    TokenViewFromJSON,
    TokenViewFromJSONTyped,
    TokenViewToJSON,
    TokenViewToJSONTyped,
} from './TokenView';

/**
 * 
 * @export
 * @interface SessionView
 */
export interface SessionView {
    /**
     * 
     * @type {AccountView}
     * @memberof SessionView
     */
    account: AccountView;
    /**
     * 
     * @type {TokenView}
     * @memberof SessionView
     */
    tokens: TokenView;
}

/**
 * Check if a given object implements the SessionView interface.
 */
export function instanceOfSessionView(value: object): value is SessionView {
    if (!('account' in value) || value['account'] === undefined) return false;
    if (!('tokens' in value) || value['tokens'] === undefined) return false;
    return true;
}

export function SessionViewFromJSON(json: any): SessionView {
    return SessionViewFromJSONTyped(json, false);
}

export function SessionViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): SessionView {
    if (json == null) {
        return json;
    }
    return {
        
        'account': AccountViewFromJSON(json['account']),
        'tokens': TokenViewFromJSON(json['tokens']),
    };
}

export function SessionViewToJSON(json: any): SessionView {
    return SessionViewToJSONTyped(json, false);
}

export function SessionViewToJSONTyped(value?: SessionView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'account': AccountViewToJSON(value['account']),
        'tokens': TokenViewToJSON(value['tokens']),
    };
}